a little refactor to the update auto reply. moved to own file
