+++ bin/db/datastore/update/updateMany.ts
@@ -1,30 +1,6 @@
 import * as chalk from 'chalk';
-import { replyInterface, updateOptionsInterface, upsertedDocument } from './interfaces';
-
-const autoReply = function ( err:Error, numAffected:number, affectedDocuments?:any, upsert?:boolean ) {
-  let replyObject:replyInterface = {};
-  if(err) console.log(chalk.red(`${err}: \n An error occurred during Update`));
-
-  if(upsert){
-    replyObject.changed = numAffected;
-    replyObject.affectedDocument = (affectedDocuments)? affectedDocuments : 0;
-    replyObject.upsert = upsert;
-  } else if(!upsert){
-    replyObject.changed = numAffected;
-    replyObject.affectedDocument = {};
-    replyObject.affectedDocuments = [];
-    if(Array.isArray(affectedDocuments)){
-      affectedDocuments.forEach(function(doc:upsertedDocument){
-        replyObject.affectedDocuments.push(doc._id);
-      });
-    } else if(affectedDocuments){
-      replyObject.affectedDocument = affectedDocuments._id;
-    }
-    replyObject.upsert = false;
-  }
-
-  console.log(replyObject);
-};
+import { autoReply } from './autoReply';
+import { updateOptionsInterface} from './interfaces';
 
 export function updateManyDocs(query: {}, updateQuery:{}, updateOptions?:updateOptionsInterface,cb?:any ){
   if(!updateOptions) updateOptions = {multi: true};
